:construction: add current_price endpoint, and add readme
