+++ config/routes.rb
@@ -4,6 +4,7 @@ Rails.application.routes.draw do
       resources :users, only: %i[index show create update] do
         resources :transactions, only: %i[index show create]
       end
+      get 'bitcoin_price', to: 'transactions#current_price_btc'
     end
   end
 end
